Builds and runs with clang plus other changes listed Default cmake to RELWITHDEBINFO for -O2 -g flags by default Clean up libexodus cmake to better use cmake find for postgres and boost Replace a lot of PUBLIC ND with ND PUBLIC to evade clang error Remove ND from all friend functions to evade clang error Refactor various code to resolve many clang warnings Add various missing header files required by clang test_sort to handle slightly different sorting order of numbers with prefixes/postfixes

+++ exodus/libexodus/exodus/term_getkey.cpp
@@ -52,7 +52,7 @@ char getkey(void) {
 	error |= tcsetattr(fd, TCSANOW, &oldAttr);
 	fcntl(fd, F_SETFL, oldFlags);
 
-	return (error ? -1 : (int)ch);
+	return (error ? -1 : ch);
 }
 
 /*
